Move common methods to GXBaseObject. Add a virtual ExecutePrivate to move executePrivateCatch from generated code to base class.
